splits encoder_advanced.hpp into fixed_size_encoder.hpp and prebuilt_codes.hpp
Previous commit already renamed encoder_advanced.hpp to fixed_size_encoder.hpp
