+++ src/LDPC4QKD/fixed_size_encoder.hpp
@@ -1,33 +1,20 @@
 //
 // Created by Adomas Baliuka on 03.12.24.
 //
-// NOTE TODO: THIS FILE SHOULD BE REMOVED once the new codes are published on GitHub
-//          right now this is somewhat in conflict with the LDPC4QKD submodule!
+// Representation of fixed-size (dimensions known at compile time) LDPC encoder (syndrome computation).
+// Supports quasi-cyclic (QC) format as input.
+//
 
 #ifndef QKD_POSTPROC_BOB_ENCODER_ADVANCED_HPP
 #define QKD_POSTPROC_BOB_ENCODER_ADVANCED_HPP
 
-#include <ranges>
 #include <cstdint>
 #include <array>
 #include <span>
-#include <random>
 #include <vector>
 #include <concepts>
 #include <sstream>
-
-#include "LDPC4QKD/rate_adaptive_code.hpp"
-
-// automatically generated LDPC codes and rate adaption arrays
-#include "LDPC4QKD/autogen_ldpc_QC.hpp"
-#include "LDPC4QKD/autogen/all_819k_codes.qccsc.hpp"
-
-#include "LDPC4QKD/autogen/rate_adaption_2x4_block_4096.hpp"
-#include "LDPC4QKD/autogen/rate_adaption_2x4_block_16384.hpp"
-#include "LDPC4QKD/autogen/rate_adaption_2x4_block_1048576.hpp"
-#include "LDPC4QKD/autogen/rate_adaption_2x6_block_6144.hpp"
-#include "LDPC4QKD/autogen/rate_adaption_2x6_block_24576.hpp"
-#include "LDPC4QKD/autogen/rate_adaption_2x6_block_1572864.hpp"
+#include <stdexcept>
 
 namespace LDPC4QKD {
     template<std::size_t N>
@@ -264,339 +251,6 @@ namespace LDPC4QKD {
         return FixedSizeEncoderQC_inst{colptr, row_idx, values};
     }
 
-    // Protograph based codes:
-    constexpr inline auto encoder_2048x6144_4663d91 = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_2048x6144_4663d91::M, AutogenLDPC_QC_2048x6144_4663d91::expansion_factor>(
-            AutogenLDPC_QC_2048x6144_4663d91::colptr, AutogenLDPC_QC_2048x6144_4663d91::row_idx,
-            AutogenLDPC_QC_2048x6144_4663d91::values);
-
-    constexpr inline auto encoder_8192x24576_71b51c1 = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_8192x24576_71b51c1::M, AutogenLDPC_QC_8192x24576_71b51c1::expansion_factor>(
-            AutogenLDPC_QC_8192x24576_71b51c1::colptr, AutogenLDPC_QC_8192x24576_71b51c1::row_idx,
-            AutogenLDPC_QC_8192x24576_71b51c1::values);
-
-    constexpr inline auto encoder_524288x1572864_4d78a9f = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_524288x1572864_4d78a9f::M, AutogenLDPC_QC_524288x1572864_4d78a9f::expansion_factor>(
-            AutogenLDPC_QC_524288x1572864_4d78a9f::colptr, AutogenLDPC_QC_524288x1572864_4d78a9f::row_idx,
-            AutogenLDPC_QC_524288x1572864_4d78a9f::values);
-
-    constexpr inline auto encoder_2048x4096_0c809c3 = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_2048x4096_0c809c3::M, AutogenLDPC_QC_2048x4096_0c809c3::expansion_factor>(
-            AutogenLDPC_QC_2048x4096_0c809c3::colptr, AutogenLDPC_QC_2048x4096_0c809c3::row_idx,
-            AutogenLDPC_QC_2048x4096_0c809c3::values);
-
-    constexpr inline auto encoder_8192x16384_3fcad37 = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_8192x16384_3fcad37::M, AutogenLDPC_QC_8192x16384_3fcad37::expansion_factor>(
-            AutogenLDPC_QC_8192x16384_3fcad37::colptr, AutogenLDPC_QC_8192x16384_3fcad37::row_idx,
-            AutogenLDPC_QC_8192x16384_3fcad37::values);
-
-    constexpr inline auto encoder_524288x1048576_9b50f98 = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_524288x1048576_9b50f98::M, AutogenLDPC_QC_524288x1048576_9b50f98::expansion_factor>(
-            AutogenLDPC_QC_524288x1048576_9b50f98::colptr, AutogenLDPC_QC_524288x1048576_9b50f98::row_idx,
-            AutogenLDPC_QC_524288x1048576_9b50f98::values);
-
-    // degree-distribution-based codes:
-    constexpr inline auto encoder_lrate_P15_block_819k = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_lrate_P15_block_819k::M, AutogenLDPC_QC_lrate_P15_block_819k::expansion_factor>(
-            AutogenLDPC_QC_lrate_P15_block_819k::colptr, AutogenLDPC_QC_lrate_P15_block_819k::row_idx,
-            AutogenLDPC_QC_lrate_P15_block_819k::values);
-
-    constexpr inline auto encoder_lrate_P1_block_819k = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_lrate_P1_block_819k::M, AutogenLDPC_QC_lrate_P1_block_819k::expansion_factor>(
-            AutogenLDPC_QC_lrate_P1_block_819k::colptr, AutogenLDPC_QC_lrate_P1_block_819k::row_idx,
-            AutogenLDPC_QC_lrate_P1_block_819k::values);
-
-    constexpr inline auto encoder_lrate_P25_block_819k = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_lrate_P25_block_819k::M, AutogenLDPC_QC_lrate_P25_block_819k::expansion_factor>(
-            AutogenLDPC_QC_lrate_P25_block_819k::colptr, AutogenLDPC_QC_lrate_P25_block_819k::row_idx,
-            AutogenLDPC_QC_lrate_P25_block_819k::values);
-
-    constexpr inline auto encoder_lrate_P2_block_819k = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_lrate_P2_block_819k::M, AutogenLDPC_QC_lrate_P2_block_819k::expansion_factor>(
-            AutogenLDPC_QC_lrate_P2_block_819k::colptr, AutogenLDPC_QC_lrate_P2_block_819k::row_idx,
-            AutogenLDPC_QC_lrate_P2_block_819k::values);
-
-    constexpr inline auto encoder_lrate_P35_block_819k = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_lrate_P35_block_819k::M, AutogenLDPC_QC_lrate_P35_block_819k::expansion_factor>(
-            AutogenLDPC_QC_lrate_P35_block_819k::colptr, AutogenLDPC_QC_lrate_P35_block_819k::row_idx,
-            AutogenLDPC_QC_lrate_P35_block_819k::values);
-
-    constexpr inline auto encoder_lrate_P3_block_819k = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_lrate_P3_block_819k::M, AutogenLDPC_QC_lrate_P3_block_819k::expansion_factor>(
-            AutogenLDPC_QC_lrate_P3_block_819k::colptr, AutogenLDPC_QC_lrate_P3_block_819k::row_idx,
-            AutogenLDPC_QC_lrate_P3_block_819k::values);
-
-    constexpr inline auto encoder_lrate_P45_block_819k = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_lrate_P45_block_819k::M, AutogenLDPC_QC_lrate_P45_block_819k::expansion_factor>(
-            AutogenLDPC_QC_lrate_P45_block_819k::colptr, AutogenLDPC_QC_lrate_P45_block_819k::row_idx,
-            AutogenLDPC_QC_lrate_P45_block_819k::values);
-
-    constexpr inline auto encoder_lrate_P4_block_819k = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_lrate_P4_block_819k::M, AutogenLDPC_QC_lrate_P4_block_819k::expansion_factor>(
-            AutogenLDPC_QC_lrate_P4_block_819k::colptr, AutogenLDPC_QC_lrate_P4_block_819k::row_idx,
-            AutogenLDPC_QC_lrate_P4_block_819k::values);
-
-    constexpr inline auto encoder_lrate_P5_block_819k = helper_create_FixedSizeEncoderQC<
-            AutogenLDPC_QC_lrate_P5_block_819k::M, AutogenLDPC_QC_lrate_P5_block_819k::expansion_factor>(
-            AutogenLDPC_QC_lrate_P5_block_819k::colptr, AutogenLDPC_QC_lrate_P5_block_819k::row_idx,
-            AutogenLDPC_QC_lrate_P5_block_819k::values);
-
-    constexpr inline std::tuple all_encoders_tuple{
-            // Protograph-based
-            encoder_2048x6144_4663d91,
-            encoder_8192x24576_71b51c1,
-            encoder_524288x1572864_4d78a9f,
-            encoder_2048x4096_0c809c3,
-            encoder_8192x16384_3fcad37,
-            encoder_524288x1048576_9b50f98,
-            // degree-distribution-based
-            encoder_lrate_P1_block_819k,
-            encoder_lrate_P15_block_819k,
-            encoder_lrate_P2_block_819k,
-            encoder_lrate_P25_block_819k,
-            encoder_lrate_P3_block_819k,
-            encoder_lrate_P35_block_819k,
-            encoder_lrate_P4_block_819k,
-            encoder_lrate_P45_block_819k,
-            encoder_lrate_P5_block_819k
-    };
-
-    namespace HelperFixedSize {
-        using Idx = std::uint32_t;
-        using Bit = std::uint8_t;
-
-        using ErrorCorrector = LDPC4QKD::RateAdaptiveCode<Idx>;
-
-        template<typename Tout, typename Tin>
-        std::vector<std::vector<Tout>> static_cast_vec_vec(std::vector<std::vector<Tin>> const &v) {
-            std::vector<std::vector<Tout>> result(v.size());
-            for (std::size_t i = 0; i < v.size(); ++i) {
-                result[i].reserve(v[i].size());
-                for (const auto &val: v[i]) {
-                    result[i].push_back(val);
-                }
-            }
-            return result;
-        }
-
-        template<typename Tout, typename Tin>
-        std::vector<Tout> static_cast_vec(std::vector<Tin> const &v) {
-            std::vector<Tout> result;
-            result.reserve(v.size());
-            for (const auto &val: v) {
-                result.push_back(val);
-            }
-            return result;
-        }
-
-        template<typename T, std::size_t N>
-        std::vector<T> arr_to_vec(std::array<T, N> a) {
-            return std::vector<T>(a.begin(), a.end());
-        }
-
-        template <typename Idx=Idx>
-        LDPC4QKD::RateAdaptiveCode<Idx> get_rate_adaptive_code(std::size_t id) {
-            switch (id) {
-                case 0: { // encoder_2048x6144_4663d91
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<0>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = static_cast_vec<Idx>(
-                            arr_to_vec(AutogenRateAdapt_2x6_block_6144::rows));
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 1: { // encoder_8192x24576_71b51c1
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<1>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = static_cast_vec<Idx>(
-                            arr_to_vec(AutogenRateAdapt_2x6_block_24576::rows));
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 2: { // encoder_524288x1572864_4d78a9f
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<2>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = static_cast_vec<Idx>(
-                            arr_to_vec(AutogenRateAdapt_2x6_block_1572864::rows));
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 3: { // encoder_2048x4096_0c809c3
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<3>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = static_cast_vec<Idx>(
-                            arr_to_vec(AutogenRateAdapt_2x4_block_4096::rows));
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 4: { // encoder_8192x16384_3fcad37
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<4>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = static_cast_vec<Idx>(
-                            arr_to_vec(AutogenRateAdapt_2x4_block_16384::rows));
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 5: { // encoder_524288x1048576_9b50f98
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<5>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = static_cast_vec<Idx>(
-                            arr_to_vec(AutogenRateAdapt_2x4_block_1048576::rows));
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 6: { // encoder 819k, lrate 0.1
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<6>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = std::vector<Idx>{}; // TODO
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 7: { // encoder 819k, lrate 0.15
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<7>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = std::vector<Idx>{}; // TODO
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 8: { // encoder 819k, lrate 0.2
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<8>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = std::vector<Idx>{}; // TODO
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 9: { // encoder 819k, lrate 0.25
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<9>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = std::vector<Idx>{}; // TODO
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 10: { // encoder 819k, lrate 0.3
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<10>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = std::vector<Idx>{}; // TODO
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 11: { // encoder 819k, lrate 0.35
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<11>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = std::vector<Idx>{}; // TODO
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 12: { // encoder 819k, lrate 0.4
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<12>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = std::vector<Idx>{}; // TODO
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 13: { // encoder 819k, lrate 0.45
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<13>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = std::vector<Idx>{}; // TODO
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                case 14: { // encoder 819k, lrate 0.5
-                    auto pos_varn = static_cast_vec_vec<Idx>(
-                            std::get<14>(all_encoders_tuple).get_pos_varn());
-                    auto rate_adapt_rows = std::vector<Idx>{}; // TODO
-                    return {pos_varn, rate_adapt_rows, 0};
-                }
-                default: {
-                    throw std::runtime_error("No code available for requested ID " + std::to_string(id));
-                }
-            }
-        }
-    }
-
-    //! Encodes the `key` using the LDPC code specified by the `code_id`. The result is the syndrome.
-    //! Note: if `code_id` known at compile time, use templated version instead!
-    //!
-    //! NOTE: Containers will be converted to a `std::span` internally.
-    //! Sizes of `key` and `result` are checked at runtime and must match exactly, otherwise an exception is thrown.
-    //!
-    //! For **containers with compile-time known sizes**, using an incorrect size may also give a COMPILE ERROR.
-    //! (something like "no matching function for call to ‘std::span<...>::span(...)").
-    //! When using such containers, the input and output sizes must exactly match ALL available codes
-    //! (which is usually impossible when there are several different codes).
-    //! If `code_id` is known at compile time, use templated version instead!
-    //! If `code_id` isn't known at compile time,
-    //! then use `std::vector` and `FixedSizeEncoder::outputSize`, `FixedSizeEncoder::inputSize`
-    //! to get the size exactly right.
-    //! Alternatively, to avoid this check completely, use `std::span`
-    //! and make sure manually that you own enough memory (otherwise you get out-of-memory access!).
-    //!
-    //! \tparam N internal implementation detail, need not use.
-    //! \param code_id integer index into tuple of codes. Make sure both sides agree on these!
-    //! \param key  Contiguous container (e.g. `std::vector`, `std::array`, `std::span`) of bits (e.g. `bool` or `uint8_t`).
-    //! \param result  Contiguous container (e.g. `std::vector`, `std::array`, `std::span`) of bits (e.g. `bool` or `uint8_t`).
-    //!                     Used to store syndrome. Must already be sized correctly for the given code!
-    template<std::size_t N = 0>
-    void encode_with(std::size_t code_id, auto const &key, auto &result) {
-        if (code_id >= std::tuple_size_v<decltype(all_encoders_tuple)>) {
-            throw std::runtime_error("Invalid code ID requested!");
-        }
-        if (N == code_id) {
-            // if/when `all_encoders_tuple` contains non-QC matrices, this needs to change!
-            // Originally, this was using `encode` instead of `encode_qc` but then it doesn't work with `vector<bool>`
-            std::get<N>(all_encoders_tuple).encode_qc(key, result);
-            return;
-        }
-
-        if constexpr (N + 1 < std::tuple_size_v<decltype(all_encoders_tuple)>) {
-            return encode_with<N + 1>(code_id, key, result);
-        }
-    }
-
-    //! Encodes the `key` using the LDPC code specified by the `code_id`.
-    //! The result is the syndrome. Note: code id must be known at compile time.
-    //! For runtime inference, use `encode_with(std::size_t code_id, auto const& key, auto &result)`
-    //!
-    //! NOTE: Containers will be converted to a `std::span` internally.
-    //! Sizes of `key` and `result` are checked at runtime and must match exactly, otherwise an exception is thrown.
-    //! For containers with compile-time known sizes, using an incorrect **size may also give a COMPILE ERROR**.
-    //! Use `get_input_size` and `get_input_size` to allocate correctly sized arrays
-    //! (to avoid this check, use `std::span` and make sure manually that you own enough memory!).
-    //!
-    //! \tparam code_id integer index into tuple of codes. Make sure both sides agree on these!
-    //! \param key Contiguous container (e.g. `std::vector`, `std::array`, `std::span`) of bits (e.g. `bool` or `uint8_t`).
-    //! \param result Contiguous container (e.g. `std::vector`, `std::array`, `std::span`) of bits (e.g. `bool` or `uint8_t`).
-    //!                 Used to store syndrome. Must already be sized correctly for the given code!
-    template<std::size_t code_id>
-    void encode_with_static(auto const &key, auto &result) {
-        std::get<code_id>(all_encoders_tuple).encode(key, result);
-    }
-
-    //! Get input size of code with given ID.
-    //!
-    //! \tparam N internal implementation detail, need not use.
-    //! \param code_id integer index into tuple of codes. Make sure both sides agree on these!
-    template<std::size_t N = 0>
-    constexpr std::size_t get_input_size(std::size_t code_id) {
-        if (code_id >= std::tuple_size_v<decltype(all_encoders_tuple)>) {
-            throw std::runtime_error("Invalid code ID requested!");
-        }
-        if (N == code_id) {
-            return std::get<N>(all_encoders_tuple).get_input_size();
-        }
-
-        if constexpr (N + 1 < std::tuple_size_v<decltype(all_encoders_tuple)>) {
-            return get_input_size<N + 1>(code_id);
-        }
-        return 0; // this should never happen
-    }
-
-    //! Get input size of code with given ID.
-    //!
-    //! \tparam N internal implementation detail, need not use.
-    //! \param code_id integer index into tuple of codes. Make sure both sides agree on these!
-    template<std::size_t N = 0>
-    constexpr std::size_t get_output_size(std::size_t code_id) {
-        if (code_id >= std::tuple_size_v<decltype(all_encoders_tuple)>) {
-            throw std::runtime_error("Invalid code ID requested!");
-        }
-        if (N == code_id) {
-            return std::get<N>(all_encoders_tuple).get_output_size();
-        }
-
-        if constexpr (N + 1 < std::tuple_size_v<decltype(all_encoders_tuple)>) {
-            return get_output_size<N + 1>(code_id);
-        }
-        return 0; // this should never happen
-    }
-
 }
-
 
 #endif //QKD_POSTPROC_BOB_ENCODER_ADVANCED_HPP
